Add PangoHbFontmap
This is a fontmap implementation for use with PangoHbFont,
which lets you add individual faces. It handles caching
and lookups and is designed to let subclasses simply provide
a populate vfunc.

Note that the load_fontset implementation here is similar in
style to the one we have in PangoFontMap: for each family
in the font description, pick the best matching face.