Tweak PangoAlignment
Remove the justify-all value, and drop the justify-last-line
feature from PangoLayout. This can be done better using
PangoLineBreaker now.

On the flip side, add a 'natural' value, since that is what
we want to fall back to for the last line in a justified
paragraph. And CoreText has this value too.

Make PANGO_ALIGN_NATURAL the default.