build: Drop references to pangowin32.rc[.in]
We are building Pango as a single library instead of 3-4 libraries, so we don't
need to build pangowin32.rc (and pangocairo.rc, which will be dropped in a
following commit).

This will also fix Meson's code configuration, as we were actually trying to
configure pangowin32.rc before the required data is configured.